Add global banner/callout bar prototype
Similar to emergency publishing, in that for all other pages on GOV.UK
it's set via a layout in static, but is custom for the homepage.

The new bar shouldn't show when emergency publishing is active, to avoid
having too many horizontal notification bars on a single page, and
emergency publishing trumps all other cases.